feat: made a lot of changes
