+++ package.json
@@ -16,6 +16,7 @@
     "@radix-ui/react-navigation-menu": "^1.2.0",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-toast": "^1.2.1",
+    "@sanity/client": "^7.8.2",
     "@sanity/image-url": "^1.0.2",
     "@sanity/react-loader": "^1.10.3",
     "@sanity/vision": "^3.48.1",
@@ -29,10 +30,11 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "framer-motion": "^11.2.12",
+    "groq": "^4.3.0",
     "lucide-react": "^0.396.0",
     "mini-svg-data-uri": "^1.4.4",
     "next": "14.2.4",
-    "next-sanity": "^9.4.2",
+    "next-sanity": "^10.0.10",
     "next-themes": "^0.3.0",
     "react": "^18",
     "react-dom": "^18",
