+++ src/app/(flagship)/alexaverse-v2/globals.css
@@ -1,0 +1,3 @@
+body {
+  background-color: #130025; /* Use the darkest color in your gradient */
+}
