--- src/app/(flagship)/alexaverse2.0/RegisterDebug/page.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import React from "react";
-import RegisterDebug from "@/components/alexaverse2.0/RegisterDebug";
-import ContactUs from "@/components/alexaverse2.0/ContactUs";
-
-const RegisterDebugPage: React.FC = () => {
-  return (
-    <main>
-      <RegisterDebug />
-      <ContactUs />
-    </main>
-  );
-};
-
-export default RegisterDebugPage;
