--- src/app/(flagship)/alexaverse2.0/RegisterHangman/page.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import React from "react";
-import RegisterHangman from "@/components/alexaverse2.0/RegisterHangman";
-import ContactUs from "@/components/alexaverse2.0/ContactUs";
-
-const RegisterHangmanPage: React.FC = () => {
-  return (
-    <main>
-      <RegisterHangman />
-      <ContactUs />
-    </main>
-  );
-};
-
-export default RegisterHangmanPage;
