--- src/app/(flagship)/alexaverse2.0/RegisterVlogit/page.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import React from "react";
-import RegisterVlogit from "@/components/alexaverse2.0/RegisterVlogit";
-import ContactUs from "@/components/alexaverse2.0/ContactUs";
-
-const RegisterVlogitPage: React.FC = () => {
-  return (
-    <main>
-      <RegisterVlogit />
-      <ContactUs />
-    </main>
-  );
-};
-
-export default RegisterVlogitPage;
