--- src/app/(flagship)/alexaverse2.0/RegisterWorkshop/page.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import React from "react";
-import RegisterWorkshop from "@/components/alexaverse2.0/RegisterWorkshop";
-import ContactUs from "@/components/alexaverse2.0/ContactUs";
-
-const RegisterWorkshopPage: React.FC = () => {
-  return (
-    <main>
-      <RegisterWorkshop />
-      <ContactUs />
-    </main>
-  );
-};
-
-export default RegisterWorkshopPage;
