--- src/app/(flagship)/alexaverse2.0/page.tsx
@@ -1,17 +1,0 @@
-"use client";
-
-import HeroSection from "@/components/alexaverse2.0/HeroSection";
-import OurEvents from "@/components/alexaverse2.0/OurEvents";
-import Sponsers from "@/components/alexaverse2.0/Sponsers"
-import ContactUs from "@/components/alexaverse2.0/ContactUs"
-
-export default function AlexaVersePage() {
-  return (
-    <main>
-      <HeroSection />
-      <OurEvents />
-      <Sponsers />
-      <ContactUs />
-    </main>
-  );
-}
