+++ src/components/global/header.tsx
@@ -78,14 +78,14 @@ const Navbar = () => {
                 </li>
               );
             })}
-            <Link href={"/alexaverse2.0"}>
+            <Link href={"/alexaverse-v2"}>
               <li
                 className={clsx(
                   "px-6 py-2 shadow-2xl drop-shadow-2xl text-white rounded-lg font-extrabold transform hover:-translate-y-1 transition duration-400"
                 )}
               >
                 <button className="px-8 py-2 rounded-full bg-gradient-to-b from-[#31B553] to-[#0AA294] text-white focus:ring-2 focus:ring-blue-400 hover:shadow-xl transition duration-200">
-                  ALEXAVERSE 2.0
+                  AlexaVerse2.0
                 </button>
                 {/* <span className="text-[#980F35]">Alexa</span>
                 <span className="text-[#AF6922]">Verse</span> */}
