+++ tailwind.config.ts
@@ -30,6 +30,7 @@ const config = {
           "var(--font-sans)", ...fontFamily.sans
         ],
         space: ["var(--font-spaceGrotesk)"],
+        moul: ["var(--font-moul)"],
         anton: ["var(--font-anton)"],
         inter: ["var(--font-inter)"],
         sairaStencilOne: [
@@ -89,6 +90,7 @@ const config = {
         gaming_button_bg: "#FF4656",
         gaming_form_bg: "#49248A",
         gaming_mobile_bg: "#E200FF",
+        dark_bg: "#130025",
       },
       borderRadius: {
         lg: "var(--radius)",
